samples/portable-sdks/python/function-chaining/requirements.txt
durabletask-azuremanaged
azure-identity
